Additional fix for staking icon
Currently the staking icon is black if coins in the wallet have weight, and the wallet has tried to stake a block in the last 30 seconds. We're extending that 30 seconds to 120 seconds to ensure the icon does not revert to white if a wallet on a slower machine cannot try to stake every 30 seconds.